split-utils * Move icon-functions to utils/icons.tsx
