+++ src/components/Map/RealtimeVehicleTag/RealtimeVehicleTag.tsx
@@ -2,10 +2,10 @@ import React from 'react'
 import { Tooltip } from '@entur/tooltip'
 import { colors } from '@entur/tokens'
 import { LegMode, TransportMode } from '@entur/sdk'
-import { getIconColor } from '../../../utils'
 import { IconColorType } from '../../../types'
 import { RealtimeVehicle } from '../../../services/realtimeVehicles/types/realtimeVehicle'
 import { useSettings } from '../../../settings/SettingsProvider'
+import { getIconColor } from '../../../utils/icon'
 import { TooltipContent } from './TooltipContent/TooltipContent'
 import './RealtimeVehicleTag.scss'
