+++ src/components/Map/StopPlaceTag/StopPlaceTag.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { colors } from '@entur/tokens'
-import { getIcon, getIconColor } from '../../../utils'
 import { IconColorType, StopPlaceWithDepartures } from '../../../types'
+import { getIcon, getIconColor } from '../../../utils/icon'
 import './StopPlaceTag.scss'
 
 const StopPlaceTag = ({ stopPlace, walkTime }: Props): JSX.Element => {
