+++ src/containers/Admin/EditTab/RealtimeDataPanel/RealTimeDataPanel.tsx
@@ -5,9 +5,10 @@ import { Loader } from '@entur/loader'
 import { Heading3, Label, Paragraph } from '@entur/typography'
 import { ExpandablePanel } from '@entur/expand'
 import { Line } from '../../../../types'
-import { getIcon, transportModeNameMapper } from '../../../../utils'
+import { transportModeNameMapper } from '../../../../utils'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { isTransport } from '../../../../utils/typeguards'
+import { getIcon } from '../../../../utils/icon'
 import { PermanentLinesPanel } from './PermanentLinesPanel'
 import './linesPanel.scss'
