+++ src/dashboards/BusStop/components/TileRows/TileRows.tsx
@@ -3,9 +3,10 @@ import { Heading3 } from '@entur/typography'
 import { DataCell, TableBody, TableRow } from '@entur/table'
 import { IconColorType, LineData } from '../../../../types'
 import { SituationModal } from '../../../../components/SituationModal/SituationModal'
-import { createTileSubLabel, getIcon, isMobileWeb } from '../../../../utils'
+import { createTileSubLabel, isMobileWeb } from '../../../../utils'
 import { SubLabelIcon } from '../SubLabelIcon/SubLabelIcon'
 import { DateRow } from '../../../../components/DateRow/DateRow'
+import { getIcon } from '../../../../utils/icon'
 import './TileRows.scss'
 
 const isMobile = isMobileWeb()
