+++ src/dashboards/Compact/BikeTile/BikeTile.tsx
@@ -6,8 +6,9 @@ import { Tile } from '../components/Tile/Tile'
 import { TileRow } from '../components/TileRow/TileRow'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { IconColorType } from '../../../types'
-import { getIconColorType, getTranslation } from '../../../utils'
+import { getTranslation } from '../../../utils'
 import { useWalkInfo, WalkInfo } from '../../../logic/use-walk-info/useWalkInfo'
+import { getIconColorType } from '../../../utils/icon'
 
 function getWalkInfoBike(
     walkInfos: WalkInfo[],
