+++ src/utils/icon.tsx
@@ -1,0 +1,143 @@
+import React from 'react'
+import { LegMode, TransportMode, TransportSubmode } from '@entur/sdk'
+import { colors } from '@entur/tokens'
+import {
+    BicycleIcon,
+    BusIcon,
+    CarferryIcon,
+    FerryIcon,
+    PlaneIcon,
+    SubwayIcon,
+    TrainIcon,
+    TramIcon,
+} from '@entur/icons'
+import { IconColorType, Theme } from '../types'
+
+function isSubModeAirportLink(subMode?: string): boolean {
+    if (!subMode) return false
+    const airportLinkTypes = ['airportLinkRail', 'airportLinkBus']
+    return airportLinkTypes.includes(subMode)
+}
+
+function isSubModeCarFerry(subMode?: string): boolean {
+    if (!subMode) return false
+    const carFerryTypes = [
+        'localCarFerry',
+        'internationalCarFerry',
+        'nationalCarFerry',
+        'regionalCarFerry',
+    ]
+
+    return carFerryTypes.includes(subMode)
+}
+
+function getIconColorType(theme: Theme | undefined): IconColorType {
+    if (!theme) return IconColorType.CONTRAST
+    const defaultThemes = [Theme.LIGHT, Theme.GREY]
+    if (defaultThemes.includes(theme)) {
+        return IconColorType.DEFAULT
+    }
+    return IconColorType.CONTRAST
+}
+
+function getIconColor(
+    type: TransportMode | LegMode | 'ferry',
+    iconColorType: IconColorType,
+    subType?: TransportSubmode,
+): string {
+    if (isSubModeAirportLink(subType)) {
+        return colors.transport[iconColorType].plane
+    }
+    switch (type) {
+        case 'bus':
+            return colors.transport[iconColorType].bus
+        case 'bicycle':
+            return colors.transport[iconColorType].mobility
+        case 'water':
+        case 'ferry':
+            return colors.transport[iconColorType].ferry
+        case 'metro':
+            return colors.transport[iconColorType].metro
+        case 'rail':
+            return colors.transport[iconColorType].train
+        case 'tram':
+            return colors.transport[iconColorType].tram
+        case 'air':
+            return colors.transport[iconColorType].plane
+        default:
+            return colors.transport[iconColorType].walk
+    }
+}
+
+type TransportIconIdentifier =
+    | 'carferry'
+    | 'bus'
+    | 'bicycle'
+    | 'ferry'
+    | 'subway'
+    | 'train'
+    | 'tram'
+    | 'plane'
+
+function getTransportIconIdentifier(
+    legMode: TransportMode | LegMode,
+    subMode?: TransportSubmode,
+): TransportIconIdentifier | null {
+    if (isSubModeCarFerry(subMode)) {
+        return 'carferry'
+    }
+
+    switch (legMode) {
+        case 'bus':
+        case 'coach':
+            return 'bus'
+        case 'bicycle':
+            return 'bicycle'
+        case 'water':
+            return 'ferry'
+        case 'metro':
+            return 'subway'
+        case 'rail':
+            return 'train'
+        case 'tram':
+            return 'tram'
+        case 'air':
+            return 'plane'
+        default:
+            return null
+    }
+}
+
+function getIcon(
+    mode: TransportMode,
+    iconColorType: IconColorType = IconColorType.CONTRAST,
+    subMode?: TransportSubmode,
+    color?: string,
+): JSX.Element | null {
+    const colorToUse = color ?? getIconColor(mode, iconColorType, subMode)
+
+    const identifier = getTransportIconIdentifier(mode, subMode)
+
+    switch (identifier) {
+        case 'bus':
+            return <BusIcon key={identifier} color={colorToUse} />
+        case 'bicycle':
+            return <BicycleIcon key={identifier} color={colorToUse} />
+        case 'carferry':
+            return <CarferryIcon key={identifier} color={colorToUse} />
+        case 'ferry':
+            return <FerryIcon key={identifier} color={colorToUse} />
+        case 'subway':
+            return <SubwayIcon key={identifier} color={colorToUse} />
+        case 'train':
+            return <TrainIcon key={identifier} color={colorToUse} />
+        case 'tram':
+            return <TramIcon key={identifier} color={colorToUse} />
+        case 'plane':
+            return <PlaneIcon key={identifier} color={colorToUse} />
+        default:
+            return null
+    }
+}
+
+export { getIconColorType, getIconColor, getTransportIconIdentifier, getIcon }
